+++ Drift/Managers/PresentationManager.swift
@@ -106,7 +106,7 @@ class PresentationManager: PresentationManagerDelegate {
         return false
     }
     
-    func showConversationList(endUserId: Int?){
+    func showConversationList(endUserId: Int64?){
         let conversationListController = ConversationListViewController.navigationController(endUserId: endUserId)
         TopController.viewController()?.present(conversationListController, animated: true, completion: nil)
     }
